requirements.txt
langchain
langchain-community
langchain-google-genai
python-dotenv
faiss-cpu
openai
tiktoken
bs4